Bug fix with a size_t
Former-commit-id: da004743ffba5331271d45ff1be1018c97e3f841

+++ libraries/reconstruction/image_rotational_spectra.cpp
@@ -118,7 +118,7 @@ void ProgMakeSpectra::run()
         // Create header
         if (first)
         {
-            int vectorSize=MULTIDIM_SIZE(rot_spt.rot_spectrum);
+            size_t vectorSize=MULTIDIM_SIZE(rot_spt.rot_spectrum);
             size_t outId=vectorHeader.addObject();
             vectorHeader.setValue(MDL_XSIZE,XSIZE(I()),outId);
             vectorHeader.setValue(MDL_YSIZE,YSIZE(I()),outId);
